+++ src/main/java/tictactoeweb/TicTacToeWeb.java
@@ -8,6 +8,12 @@ import org.cobspec.controller.FileSystemController;
 import org.cobspec.controller.OptionsController;
 import org.flint.Application;
 
+import flint.cors.CorsOptions;
+import flint.cors.CorsMiddleware;
+import tictactoeweb.schema.SchemaStore;
+import tictactoeweb.tictactoe.TicTacToeController;
+import tictactoeweb.schema.SchemaController;
+
 public class TicTacToeWeb {
     private static final int DEFAULT_PORT = 5000;
 
@@ -17,7 +23,7 @@ public class TicTacToeWeb {
             .exposeHeaders("Link,Location")
             .build();
 
-        final ApplicationOptions options = new ApplicationOptions.Builder()
+        final TicTacToeWebOptions options = new TicTacToeWebOptions.Builder()
             .dataPath(Paths.get("."))
             .schemaPath(Paths.get("src/main"))
             .webPath(Paths.get("web"))
@@ -28,7 +34,7 @@ public class TicTacToeWeb {
         build(options).run(port);
     }
 
-    static Application build(final ApplicationOptions options) {
+    static Application build(final TicTacToeWebOptions options) {
         final Application app = new Application();
 
         ticTacToeApi(app, options.getDataPath(), options.getSchemaPath());
