+++ src/main/schema/play.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "object",
+  "properties": {
+    "position": {
+      "enum": [
+        "topLeft", "topMiddle", "topRight",
+        "middleLeft", "center", "middleRight",
+        "bottomLeft", "bottomMiddle", "bottomRight"
+      ]
+    }
+  },
+  "required": ["position"]
+}
