+++ src/test/java/flint/cors/CorsMiddlewareTest.java
@@ -1,4 +1,4 @@
-package tictactoeweb;
+package flint.cors;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
