+++ src/test/java/json/JsonTest.java
@@ -1,4 +1,4 @@
-package tictactoeweb;
+package json;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
