+++ src/test/java/tictactoeweb/schema/SchemaStoreTest.java
@@ -1,4 +1,4 @@
-package tictactoeweb;
+package tictactoeweb.schema;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
